+++ prisma/schema.prisma
@@ -21,50 +21,6 @@ enum Visibility {
   FRIENDS
 }
 
-model Account {
-  id                 String    @id @default(cuid())
-  userId             String
-  providerType       String
-  providerId         String
-  providerAccountId  String
-  refreshToken       String?
-  accessToken        String?
-  accessTokenExpires DateTime?
-  createdAt          DateTime  @default(now())
-  updatedAt          DateTime  @updatedAt
-
-  user User @relation(fields: [userId], references: [id])
-
-  @@unique([providerId, providerAccountId])
-}
-
-model User {
-  id            String   @id @default(cuid())
-  avatar        String?
-  email         String   @unique
-  emailVerified Boolean  @default(false)
-  username      String   @unique
-  password      String
-  firstName     String?
-  isActive      Boolean  @default(true)
-  role          Role     @default(USER)
-  createdAt     DateTime @default(now())
-  updatedAt     DateTime @updatedAt
-
-  shoppingList      ShoppingList?
-  createdRecipes    Recipe[]
-  accounts          Account[]
-  verificationToken VerificationToken?
-}
-
-model VerificationToken {
-  id      String   @id @default(cuid())
-  token   String   @unique
-  expires DateTime
-  userId  String   @unique
-  user    User     @relation(fields: [userId], references: [id])
-}
-
 enum Unit {
   GRAM
   MILLIGRAM
@@ -80,51 +36,101 @@ enum Unit {
   PINCH
 }
 
+model Account {
+  id                 String    @id @default(cuid(2))
+  providerType       String
+  providerId         String
+  providerAccountId  String
+  refreshToken       String?
+  accessToken        String?
+  accessTokenExpires DateTime?
+  createdAt          DateTime  @default(now())
+  updatedAt          DateTime  @updatedAt
+
+  userId String
+  user   User   @relation(fields: [userId], references: [id])
+
+  @@unique([providerId, providerAccountId])
+}
+
+model User {
+  id            String   @id @default(cuid(2))
+  avatar        String?
+  email         String   @unique
+  emailVerified Boolean  @default(false)
+  username      String   @unique
+  password      String
+  firstName     String?
+  isActive      Boolean  @default(true)
+  createdAt     DateTime @default(now())
+  updatedAt     DateTime @updatedAt
+  role          Role     @default(USER)
+
+  verificationToken VerificationToken?
+  shoppingList      ShoppingList?
+  createdRecipes    Recipe[]
+  accounts          Account[]
+}
+
+model VerificationToken {
+  id      String   @id @default(cuid(2))
+  token   String   @unique
+  expires DateTime
+
+  userId String @unique
+  user   User   @relation(fields: [userId], references: [id])
+}
+
 model Ingredient {
-  id       Int     @id @default(autoincrement())
+  id       Int    @id @default(autoincrement())
   name     String
   quantity Float
   unit     Unit
-  Recipe   Recipe? @relation(fields: [recipeId], references: [id])
-  recipeId Int?
+
+  recipeId String
+  Recipe   Recipe @relation(fields: [recipeId], references: [id])
 }
 
 model Recipe {
-  id               Int          @id @default(autoincrement())
-  name             String
-  slug             String       @unique
-  description      String
-  steps            String[]
-  ingredients      Ingredient[]
-  visibility       Visibility   @default(PRIVATE)
-  images           String[]
-  recipeCategoryId Int
-  authorId         String
-  createdAt        DateTime     @default(now())
-  updatedAt        DateTime     @updatedAt
+  id          String     @id @default(cuid(2))
+  name        String
+  slug        String
+  description String
+  steps       String[]
+  images      String[]
+  createdAt   DateTime   @default(now())
+  updatedAt   DateTime   @updatedAt
+  visibility  Visibility @default(PRIVATE)
 
-  category RecipeCategory? @relation(fields: [recipeCategoryId], references: [id])
-  author   User            @relation(fields: [authorId], references: [id])
-  tag      Tag[]
+  tag         Tag[]
+  ingredients Ingredient[]
+  categoryId  Int            @map("recipeCategoryId")
+  category    RecipeCategory @relation(fields: [categoryId], references: [id])
+  authorId    String
+  author      User           @relation(fields: [authorId], references: [id])
+
+  @@unique([slug, id])
 }
 
 model RecipeCategory {
-  id      Int      @id @default(autoincrement())
-  name    String
-  slug    String   @unique
+  id   Int    @id @default(autoincrement())
+  name String
+  slug String @unique
+
   recipes Recipe[]
 }
 
 model Tag {
-  id      Int      @id @default(autoincrement())
-  name    String   @unique
+  id   Int    @id @default(autoincrement())
+  name String @unique
+
   recipes Recipe[]
 }
 
 model ShoppingList {
-  id     Int    @id @default(autoincrement())
-  userId String @unique
-
-  user        User     @relation(fields: [userId], references: [id])
+  id          Int      @id @default(autoincrement())
   ingredients String[]
+
+  userId String @unique
+  user   User   @relation(fields: [userId], references: [id])
 }
